refactor(frontend): remove old login page
